refactor: make pipeline fully project-aware and clean legacy scripts

+++ 08_abricate.sh
@@ -4,10 +4,19 @@ set -euo pipefail
 eval "$(conda shell.bash hook)"
 conda activate 07_abricate
 
-BASE="/data/wgs_assembly/hybrid_genome_assembly_guide"
-ASM="$BASE/05_genome_assembly/03_hybrid/assembly.fasta"
-PLASM="$BASE/08_plassembler/plassembler_plasmids.fasta"
-OUT="$BASE/09_abricate"
+# ===============================
+# INPUT ARGUMENT
+# ===============================
+RESULTS=$1
+
+if [ -z "$RESULTS" ]; then
+  echo "Usage: 08_abricate.sh <RESULTS_DIR>"
+  exit 1
+fi
+
+ASM="$RESULTS/assembly/03_hybrid/assembly.fasta"
+PLASM="$RESULTS/plasmids/plassembler_plasmids.fasta"
+OUT="$RESULTS/amr"
 
 # clean old results
 rm -rf "$OUT"
@@ -20,6 +29,7 @@ for db in "${DBS[@]}"; do
   abricate --db "$db" "$ASM" > \
     "$OUT/assembly_level/abricate_${db}_assembly.tsv"
 done
+
 abricate --summary "$OUT"/assembly_level/abricate_*_assembly.tsv > \
   "$OUT/assembly_level/abricate_assembly_summary.tsv"
 
@@ -28,6 +38,7 @@ for db in "${DBS[@]}"; do
   abricate --db "$db" "$PLASM" > \
     "$OUT/plasmid_level/abricate_${db}_plasmid.tsv"
 done
+
 abricate --summary "$OUT"/plasmid_level/abricate_*_plasmid.tsv > \
   "$OUT/plasmid_level/abricate_plasmid_summary.tsv"
